fix: remove route validation
