+++ lib/route.js
@@ -122,8 +122,7 @@ function buildRouting (options) {
     addConstraintStrategy,
     hasConstraintStrategy,
     isAsyncConstraint,
-    findRoute,
-    validateRouteParams
+    findRoute
   }
 
   function addConstraintStrategy (strategy) {
@@ -180,22 +179,6 @@ function buildRouting (options) {
       options.url || '',
       options.constraints
     )
-  }
-
-  function validateRouteParams (options) {
-    const { method, url, constraints } = options
-    const route = findRoute({ method, url: url || '', constraints })
-    if (route === null) {
-      return false
-    }
-
-    const schemaParams = new Set(Object.keys(route.store.schema.params))
-    const params = new Set(Object.keys(route.params))
-    if (schemaParams.size !== params.size) {
-      return false
-    }
-
-    return [...schemaParams].every(param => params.has(param))
   }
 
   /**
